+++ database.js
@@ -1,0 +1,8 @@
+
+function initialize_tables() {
+    
+}
+
+
+
+
